Support working with multiple cameras Added an ability to load Advanced Mode JSON files Added dynamic params for SR300, RS415 and RS435
