+++ .gitignore
@@ -1,3 +1,5 @@
 *.bin
 *.ppm
 *.bag
+*.pyc
+*.ini
