+++ realsense_ros_camera/cfg/base_d400_params.cfg
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+
+PACKAGE="realsense_ros_camera"
+
+from dynamic_reconfigure.parameter_generator_catkin import *
+
+import base_d400_params
+
+gen = ParameterGenerator()
+
+base_d400_params.add_base_params(gen, "base_")
+
+exit(gen.generate(PACKAGE, "realsense_ros_camera", "base_d400_params"))
+
