+++ realsense_ros_camera/launch/demo_pointcloud.launch
@@ -1,23 +1,27 @@
 <launch>
-  <arg name="manager" default="realsense_ros_camera_manager"/>
+  <arg name="serial_no"    default=""/>
+  <arg name="json_file_path"  default=""/>
 
-  <include file="$(find realsense_ros_camera)/launch/rs_camera.launch">
-    <arg name="manager"           value="$(arg manager)"/>
-    <arg name="depth_width"       value="640"/>
-    <arg name="depth_height"      value="480"/>
-    <arg name="depth_fps"         value="30"/>
-    <arg name="color_width"       value="640"/>
-    <arg name="color_height"      value="480"/>
-    <arg name="color_fps"         value="30"/>
-    <arg name="enable_depth"      value="true"/>
-    <arg name="enable_color"      value="true"/>
-    <arg name="enable_infra1"     value="false"/>
-    <arg name="enable_infra2"     value="false"/>
-    <arg name="enable_fisheye"    value="false"/>
-    <arg name="enable_imu"        value="false"/>
-    <arg name="enable_pointcloud" value="true"/>
-    <arg name="enable_sync"       value="true"/>
-  </include>
+  <group ns="camera">
+    <include file="$(find realsense_ros_camera)/launch/includes/nodelet.launch.xml">
+      <arg name="serial_no"         value="$(arg serial_no)"/>
+      <arg name="json_file_path"    value="$(arg json_file_path)"/>
+      <arg name="depth_width"       value="640"/>
+      <arg name="depth_height"      value="480"/>
+      <arg name="depth_fps"         value="30"/>
+      <arg name="color_width"       value="640"/>
+      <arg name="color_height"      value="480"/>
+      <arg name="color_fps"         value="30"/>
+      <arg name="enable_depth"      value="true"/>
+      <arg name="enable_color"      value="true"/>
+      <arg name="enable_infra1"     value="false"/>
+      <arg name="enable_infra2"     value="false"/>
+      <arg name="enable_fisheye"    value="false"/>
+      <arg name="enable_imu"        value="false"/>
+      <arg name="enable_pointcloud" value="true"/>
+      <arg name="enable_sync"       value="true"/>
+    </include>
 
-  <node name="rviz" pkg="rviz" type="rviz" args="-d $(find realsense_ros_camera)/rviz/pointcloud.rviz" required="true" />
+    <node name="rviz" pkg="rviz" type="rviz" args="-d $(find realsense_ros_camera)/rviz/pointcloud.rviz" required="true" />
+  </group>
 </launch>
