+++ realsense_ros_camera/package.xml
@@ -15,6 +15,7 @@
   <build_depend>cv_bridge</build_depend>
   <build_depend>image_transport</build_depend>
   <build_depend>tf</build_depend>
+  <build_depend>dynamic_reconfigure</build_depend>
   <build_depend>librealsense2</build_depend>
   <run_depend>image_transport</run_depend>
   <run_depend>cv_bridge</run_depend>
@@ -25,6 +26,7 @@
   <run_depend>std_msgs</run_depend>
   <run_depend>message_runtime</run_depend>
   <run_depend>tf</run_depend>
+  <run_depend>dynamic_reconfigure</run_depend>
   <run_depend>librealsense2</run_depend>
   <export>
     <nodelet plugin="${prefix}/nodelet_plugins.xml" />
